add extension API function pointer checks
This PR adds checks to ensure that any function pointer in the
ICD dispatch table is non-NULL before calling it.  If a function
pointer is NULL then CL_INVALID_VALUE is returned.  This
improves robustness for applications that assume they can safely
call a non-NULL function pointer returned by
clGetExtensionFunctionAddress or clGetExtensionFunctionAddressForPlatform.